Fix unmotivated change which causes mismatch between function declaration and definition.
Review URL: http://codereview.chromium.org/113640

git-svn-id: http://v8.googlecode.com/svn/branches/bleeding_edge@2018 ce2b1a6d-e550-0410-aec6-3dcde31c8c00

+++ src/ia32/assembler-ia32.cc
@@ -354,7 +354,7 @@ void Assembler::GetCode(CodeDesc* desc) {
 }
 
 
-void Assembler::Align(intptr_t m) {
+void Assembler::Align(int m) {
   ASSERT(IsPowerOf2(m));
   while ((pc_offset() & (m - 1)) != 0) {
     nop();
